03/01/18 -1 Commit
Found the flavour text issue
the Flavour string was never cleared if the text was too big so it kept getting longer and longer (each recursive loop added another iteration of the same text) which made the text taller and taller and hence font got smaller and smaller to accomodate.
Just one more minor thing, gave the flavour text 2 more pixels vertical room to draw in (so they only have 1 pixel wide padding now but that should be ok).