05/01/19-1 Commit
Because of the issues with having a comma diliminated string for the abilities, changed this to a more common pipe deliminator '|'. And then while I was at it, changed it to abilities can have empty text boxes (this crashed before hand). This solves the crash issue but also means passive abilities without a trigger can be written without splitting things between multiple lines.