06/01/19 Commit
Finally changed over to a list based set viewer which allows for the cards to always be sorted in order of set code . The issues mentioned previously were because i was using selected index not selected item (the index on the listview and in the list werent necessarily the same since the listview was reorganised) also made code more efficient as anywhere i would have to select from the listview of a set instead of a pair of casts (one to a listviewitem and then one to the cardoverview I needed) since the items were classoverviews already I could just do one cast. Also fixed the bug with trying to delete sets freezing the program, was because I had a pair of SQL queries firing at nearly the same time so the database was getting deadlocked, all my other queries had a relativly decent time between each of them so it wasnt an issue there.