15/04/19 Commit
Finally finished factorising most of the code down into functions and cleaning up the card rendering code with these new functions. Replaced all the ability icons and textbox layout . Tinkered with the primary names bounding box so it accounts for the absurd padding the glyphs have on the top and bottom so the text can be bigger and thus more readable. Added a spot of text for acknowledging the artist who made the card artwork, opposite the set code